:recycle: Prevent a player from getting targeted by his mission

+++ src/modules/mission/mission.service.ts
@@ -33,13 +33,27 @@ export class MissionService {
     player: PlayerModel,
     updateMissionContent: string,
   ): Promise<MissionModel> {
-    await this.checkMissionBelongToPlayer(missionId, player);
+    const canDeleteMission = await this.checkMissionBelongToPlayer(
+      missionId,
+      player,
+    );
+
+    if (!canDeleteMission) {
+      throw new NotFoundException({ key: 'player.MISSION.NOT_FOUND' });
+    }
 
     return this.missionRepo.updateMission(missionId, updateMissionContent);
   }
 
   async deleteMission(player: PlayerModel, missionId): Promise<void> {
-    await this.checkMissionBelongToPlayer(missionId, player);
+    const canDeleteMission = await this.checkMissionBelongToPlayer(
+      missionId,
+      player,
+    );
+
+    if (!canDeleteMission) {
+      throw new NotFoundException({ key: 'player.MISSION.NOT_FOUND' });
+    }
 
     return this.missionRepo.deleteMission(missionId);
   }
@@ -47,16 +61,12 @@ export class MissionService {
   async checkMissionBelongToPlayer(
     missionId: number,
     player: PlayerModel,
-  ): Promise<void> {
+  ): Promise<boolean> {
     const playerMissions = await this.getMissionsByPlayer(player);
 
-    const mission = playerMissions.find(
+    return !!playerMissions.find(
       (playerMission) => playerMission?.id === missionId,
     );
-
-    if (!mission) {
-      throw new NotFoundException({ key: 'player.MISSION.NOT_FOUND' });
-    }
   }
 
   async clearPlayerMissions(player: PlayerModel): Promise<void> {
